fix: multiple small issues fix (#35)
* fix: print dirty version if version is invalid

When someone tries to build minutes without providing the necessary
build flags, using version flag will result in an error. This commit
resolves the issue by printing "dirty build" in these cases.

* fix: ensure config file is not necessary

The config file was necessary to use minutes, though it should not be.
In fact, minutes should be able to run without any configuration file.

* docs: add more generic source and target options

* test: remove race condition detection

Timewarrior tests has race conditions and that was introduced keeping
this in mind. Since it was a conscious decision, it makes no sense to
check for race conditions.

Signed-off-by: Gabor Boros <gabor.brs@gmail.com>