Rework tag parsing (#36)
* refactor(tag-parsing)!: remove tags-as-tasks flag

To determine if tags should be treated as flags should happen based on the value of the corresponding regex. If the regex is empty (""), the condition should be false, otherwise, we should treat tags as tasks.

BREAKING CHANGE: This refactoring corrects a wrong assumption in the timewarrior source, that used tags-as-tasks flag. From now on, the timewarrior client will split entries as the other sources if the split condition matches.

* test: use `ElementsMatch` over `Equal`
* refactor: move tags-as-tasks in-house to FetchOpts
* docs: fix comment wording
* refactor: unify how regex is checked
* chore(changelog): update changelog